ttnn/conv: split test_conv2d.py into 3 files, add use_module_device
test_conv2d.py had two incompatible device_params values across its
test functions (l1_small_size: 16384 and 32768), making a single
module-level device impossible.

Refactor into three files:
- test_conv2d_common.py: shared imports/aliases (run_conv, HS/WS/BS,
  SliceHeight, SliceWidth)
- test_conv2d_l1_16384.py: test_conv_features (192 cases) +
  test_conv_dram_config (3 cases), pytestmark use_module_device 16384
- test_conv2d_l1_32768.py: test_conv_dram (14 cases),
  pytestmark use_module_device 32768

Also add pytestmark = pytest.mark.use_module_device({"l1_small_size":
16384}) to test_fold_op.py (306 test cases, highest-value target).

Combined savings: ~509 device open/close cycles per CI run for the
conv unit test suite.
